chore(refactor): update filenames
